+++ middleware.ts
@@ -4,14 +4,18 @@ export function middleware(req: NextRequest) {
   const sessionCookie = req.cookies.get('session')?.value
   const [id, role] = sessionCookie ? sessionCookie.split(':') : [null, null]
 
-  const protectedPaths = ['/dashboard', '/admin']
+  const protectedPaths = ['/', '/dashboard', '/admin']
 
-  if (protectedPaths.some(path => req.nextUrl.pathname.startsWith(path))) {
+  if (
+    protectedPaths.some(path =>
+      req.nextUrl.pathname === path || req.nextUrl.pathname.startsWith(`${path}/`)
+    )
+  ) {
     if (!id) {
-      return NextResponse.redirect(new URL('/', req.url))
+      return NextResponse.redirect(new URL('/login', req.url))
     }
     if (req.nextUrl.pathname.startsWith('/admin') && role !== 'admin') {
-      return NextResponse.redirect(new URL('/dashboard', req.url))
+      return NextResponse.redirect(new URL('/', req.url))
     }
   }
 
@@ -19,5 +23,5 @@ export function middleware(req: NextRequest) {
 }
 
 export const config = {
-  matcher: ['/dashboard/:path*', '/admin/:path*'],
+  matcher: ['/', '/dashboard/:path*', '/admin/:path*'],
 }
